+++ cli/transit-cli/CHANGELOG.md
@@ -9,7 +9,7 @@
 ### Patch Changes
 
 - Updated dependencies [298063f]
-  - @transit/scanner@1.0.0
-  - @transit/codegen@1.0.0
+  - @sabeeirsharrma/scanner@1.0.0
+  - @sabeeirsharrma/codegen@1.0.0
   - transit@1.0.0
-  - @transit/schema@1.0.0
+  - @sabeeirsharrma/schema@1.0.0