+++ packages/transit-codegen/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @transit/codegen
+# @sabeeirsharrma/codegen
 
 ## 1.0.0
 
@@ -9,4 +9,4 @@
 ### Patch Changes
 
 - Updated dependencies [298063f]
-  - @transit/schema@1.0.0
+  - @sabeeirsharrma/schema@1.0.0